update :: 데이터 타입 변경

+++ src/main/java/com/mindway/server/v2/domain/goal/presentation/dto/request/GoalAddRequestDto.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class GoalAddRequestDto {
     @NotNull
-    private Long goal_count;
+    private Integer goal_count;
 }
